Move Redis drivers to new namespace
